// src/app/models/usuario.model.ts
export class UsuarioModel {
  id_usuario: number;
  nombre: string;
  ap_paterno: string;
  ap_materno: string;
  email: string;
  password: string;
  password2: string;
  cargo_id_cargo: number;
  division_id_division: number;
  rol_id_rol: number;
  token:string;
  usuario: string;
  id_ticket: number;
  ci: string;
}